refactor: Использовать интерфейсы
